requirements.txt
coverage==5.5
dj-rest-auth==2.1.4
Django==3.2
django-admin-interface==0.16.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.2
django-lifecycle==0.9.1
django-redis==4.12.1
djangorestframework==3.12.4
drf-spectacular==0.15.1
factory-boy==3.2.0
gevent==21.1.2
gunicorn==20.1.0
pandas==1.2.4
psycopg2-binary==2.8.6
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
redis==3.5.3
requests==2.25.1
sentry-sdk==1.0.0
wca-django-allauth==1.0.1
Werkzeug==1.0.1
whitenoise==5.2.0
